% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run an R package}
\usage{
run(path, ...)
}
\arguments{
\item{path}{path to a package tarball}

\item{...}{other arguments (currently ignored)}
}
\description{
Installs an R package provided as a tarball, installs its dependencies,
and runs the package's `main()` function.
}
\examples{
\dontrun{
run('myapp_0.1.0.tar.gz')
}
}
